refactor: 좌표 변환 서비스 패키지 이동
